Product representation to fill Orders (#18)
* Added product class
