+++ .gitignore
@@ -1,0 +1,3 @@
+*.classpath
+*.project
+/bin/
