--- Order.java
@@ -1,31 +1,0 @@
-
-
-/**
- * @author tom
- *
- */
-public class Order {
-	private  int customerNumber;
-	private  int orderNumber;
-	private String address;
-	public int getCustomerNumber() {
-		return customerNumber;
-	}
-	public void setCustomerNumber(int customerNumber) {
-		this.customerNumber = customerNumber;
-	}
-	public int getOrderNumber() {
-		return orderNumber;
-	}
-	public void setOrderNumber(int orderNumber) {
-		this.orderNumber = orderNumber;
-	}
-	public String getAddress() {
-		return address;
-	}
-	public void setAddress(String address) {
-		this.address = address;
-	}
-	
-
-}
